# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "EcoRecycle API",
    "description": "Dokumentasi REST API untuk EcoRecycle, platform manajemen penjemputan dan daur ulang sampah elektronik.",
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "Current EcoRecycle server"
    }
  ],
...
  "paths": {
    "/api/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Login pengguna",
        "description": "Mengembalikan signed token HMAC untuk request API berikutnya.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              },
              "example": {
                "email": "user@ecorecycle.com",
                "password": "password123"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login berhasil",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/api/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "Registrasi akun",
        "description": "Membuat akun baru. Role yang valid: user, collector, admin. Jika role kosong atau tidak valid, otomatis menjadi user.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              },
              "example": {
                "name": "Eco Warrior",
                "email": "newuser@ecorecycle.com",
                "password": "password123",
                "role": "user"
              }
            }
          }
        },
        "responses": {
          "201": {
            "$ref": "#/components/responses/Created"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/api/user/profile": {
      "get": {
        "tags": ["Auth"],
        "summary": "Ambil profil pengguna",
        "description": "Mengambil profil pengguna aktif beserta data payout reward.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Profil berhasil diambil",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "post": {
        "tags": ["Auth"],
        "summary": "Update profil dan data payout",
        "description": "Menyimpan nama, kontak, alamat default, dan tujuan payout reward pengguna.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileUpdateRequest"
              },
              "example": {
                "name": "Eco Warrior",
                "phone": "08123456789",
                "address": "Jl. Asia Afrika No. 1, Bandung",
                "payout_method": "ewallet",
                "payout_account_name": "Eco Warrior",
                "payout_account_number": "08123456789"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profil berhasil diperbarui",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/api/users/collectors": {
      "get": {
        "tags": ["Auth"],
        "summary": "Ambil daftar Eco Collector",
        "description": "Mengambil daftar akun collector untuk penugasan pickup oleh admin.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Daftar kolektor berhasil diambil",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectorListResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/api/ecorecycle/request_pickup": {
      "post": {
        "tags": ["Pickup"],
        "summary": "Ajukan penjemputan e-waste",
        "description": "Membuat permohonan penjemputan baru. Hanya dapat digunakan oleh role user atau Eco Warrior.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/RequestPickupForm"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestPickupJson"
              },
              "example": {
                "item_description": "2 HP rusak dan 1 tablet mati",
                "pickup_address": "Jl. Asia Afrika No. 1, Bandung",
                "contact_phone": "08123456789",
                "weight_kg": 3,
                "category": "Small Gadgets"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Permohonan penjemputan berhasil dibuat",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestPickupResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/api/ecorecycle/update_pickup": {
      "post": {
        "tags": ["Pickup"],
        "summary": "Edit pengajuan pickup pending",
        "description": "Eco Warrior dapat mengubah data pengajuan selama status masih pending atau belum diambil kolektor.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePickupRequest"
              },
              "example": {
                "tracking_number": "ECR-20260715-12345",
                "item_description": "2 HP rusak dan 1 tablet mati",
                "pickup_address": "Jl. Asia Afrika No. 1, Bandung",
                "contact_phone": "08123456789",
                "weight_kg": 3.2,
                "category": "Small Gadgets"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/Ok"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/api/ecorecycle/cancel_pickup": {
      "post": {
        "tags": ["Pickup"],
        "summary": "Batalkan pengajuan pickup pending",
        "description": "Eco Warrior dapat membatalkan pengajuan selama status masih pending atau belum diambil kolektor.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelPickupRequest"
              },
              "example": {
                "tracking_number": "ECR-20260715-12345",
                "reason": "Alamat pickup berubah, akan membuat pengajuan baru."
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/Ok"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/api/ecorecycle/list_pickups": {
      "get": {
        "tags": ["Pickup"],
        "summary": "Ambil daftar penjemputan",
        "description": "Mengambil daftar pickup berdasarkan role dan tipe query. type=user untuk Eco Warrior, type=collector untuk Eco Collector, type=all untuk Admin.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["user", "collector", "all"],
              "default": "all"
            },
            "example": "user"
          }
        ],
        "responses": {
          "200": {
            "description": "Daftar penjemputan berhasil diambil",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PickupListResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/api/ecorecycle/assign_collector": {
      "post": {
        "tags": ["Pickup"],
        "summary": "Ambil atau tetapkan tugas kolektor",
        "description": "Mengubah status pickup dari pending menjadi pickup. Role collector mengambil tugas untuk dirinya sendiri, role admin dapat mengisi collector_id.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignCollectorRequest"
              },
              "example": {
                "tracking_number": "ECR-20260709-12345"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tugas penjemputan berhasil diambil",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssignCollectorResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/api/ecorecycle/pickup_status": {
      "get": {
        "tags": ["Tracking"],
        "summary": "Lacak pickup atau ambil statistik",
        "description": "Gunakan tracking_number untuk detail pelacakan, atau type=stats untuk statistik dashboard admin.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "tracking_number",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "ECR-20260709-12345"
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["stats"]
            },
            "example": "stats"
          }
        ],
        "responses": {
          "200": {
            "description": "Data pelacakan atau statistik berhasil diambil",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/PickupDetailResponse"
                    },
                    {
                      "$ref": "#/components/schemas/StatsResponse"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "post": {
        "tags": ["Tracking"],
        "summary": "Perbarui status pickup",
        "description": "Mengubah status pickup operasional. Transisi valid: pickup ke transit, lalu transit ke arrived. Hanya collector pemilik tugas atau admin.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePickupStatusRequest"
              },
              "example": {
                "tracking_number": "ECR-20260709-12345",
                "status": "transit",
                "location": "Dalam perjalanan ke hub",
                "notes": "Barang sudah dijemput dari alamat donor"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Status pickup berhasil diperbarui",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdatePickupStatusResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/api/ecorecycle/estimate_reward": {
      "post": {
        "tags": ["Reward"],
        "summary": "Hitung estimasi eco-reward",
        "description": "Menghitung reward dan estimasi CO2 terselamatkan berdasarkan kategori dan berat.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EstimateRewardRequest"
              },
              "example": {
                "category": "Small Gadgets",
                "weight_kg": 3
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Estimasi reward berhasil dihitung",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EstimateRewardResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/api/ecorecycle/process_payout": {
      "post": {
        "tags": ["Reward"],
        "summary": "Proses payout reward",
        "description": "Memproses payout untuk pickup yang sudah arrived. Hanya role admin.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProcessPayoutRequest"
              },
              "example": {
                "pickup_id": 1
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Payout berhasil diproses",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessPayoutResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/api/ecorecycle/verify_pickup": {
      "post": {
        "tags": ["Reward"],
        "summary": "Verifikasi berat final pickup",
        "description": "Admin memverifikasi berat final hasil timbangan hub sebelum payout reward diproses.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyPickupRequest"
              },
              "example": {
                "pickup_id": 2,
                "final_weight_kg": 4.1,
                "verification_notes": "Berat final berdasarkan timbangan hub."
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Berat final berhasil diverifikasi",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerifyPickupResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/api/ecorecycle/announcements": {
      "get": {
        "tags": ["Announcements"],
        "summary": "Ambil pengumuman portal",
        "description": "Mengambil pengumuman aktif untuk user/collector. Admin melihat semua pengumuman termasuk yang nonaktif.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Pengumuman berhasil diambil",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnnouncementListResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "post": {
        "tags": ["Announcements"],
        "summary": "Buat atau ubah status pengumuman",
        "description": "Admin dapat membuat pengumuman baru. Untuk toggle status, kirim action=toggle, id, dan is_active.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/CreateAnnouncementRequest"
                  },
                  {
                    "$ref": "#/components/schemas/ToggleAnnouncementRequest"
                  }
                ]
              },
              "examples": {
                "create": {
                  "summary": "Buat pengumuman",
                  "value": {
                    "title": "Jadwal pickup minggu ini",
                    "message": "Penjemputan wilayah Bandung Timur diprioritaskan pada hari Jumat.",
                    "target_role": "collector"
                  }
                },
                "toggle": {
                  "summary": "Aktif/nonaktifkan pengumuman",
                  "value": {
                    "action": "toggle",
                    "id": 1,
                    "is_active": false
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/Ok"
          },
          "201": {
            "$ref": "#/components/responses/Created"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "delete": {
        "tags": ["Announcements"],
        "summary": "Hapus pengumuman",
        "description": "Menghapus pengumuman portal berdasarkan id. Hanya role admin.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Ok"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "HMAC-SHA256 signed token"
      }
    },
    "responses": {
      "Ok": {
        "description": "Request berhasil",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiResponse"
            }
          }
        }
      },
      "Created": {
        "description": "Data berhasil dibuat",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiResponse"
            }
          }
        }
      },
      "BadRequest": {
        "description": "Input tidak valid",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiResponse"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Token tidak ada atau tidak valid",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiResponse"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Role pengguna tidak memiliki akses",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiResponse"
            }
          }
        }
      },
      "NotFound": {
        "description": "Data tidak ditemukan",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiResponse"
            }
          }
        }
      },
      "MethodNotAllowed": {
        "description": "Metode HTTP tidak diizinkan",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiResponse"
            }
          }
        }
      },
      "ServerError": {
        "description": "Kesalahan server internal",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiResponse"
            }
          }
        }
      }
    },
    "schemas": {
...
      "LoginRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "format": "password",
            "minLength": 6
          }
        }
      },
      "RegisterRequest": {
        "type": "object",
        "required": ["name", "email", "password"],
        "properties": {
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "format": "password",
            "minLength": 6
          },
          "role": {
            "type": "string",
            "enum": ["user", "collector", "admin"],
            "default": "user"
          }
        }
      },
...
      "ProfileUpdateRequest": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "payout_method": {
            "type": "string",
            "enum": ["bank", "ewallet", "cash"]
          },
          "payout_account_name": {
            "type": "string"
          },
          "payout_account_number": {
            "type": "string"
          }
        }
      },
      "ProfileResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiResponse"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "$ref": "#/components/schemas/UserProfile"
              }
            }
          }
        ]
      },
      "CollectorListResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiResponse"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/UserProfile"
                }
              }
            }
          }
        ]
      },
      "LoginResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiResponse"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "integer"
                  },
                  "name": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "role": {
                    "type": "string",
                    "enum": ["user", "collector", "admin"]
                  },
                  "token": {
                    "type": "string"
                  }
                }
              }
            }
          }
        ]
      },
...
    }
  }
}